update files and fix errors

+++ 0x15-file_io/0-read_textfile.c
@@ -10,7 +10,7 @@
 
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	size_t o, r, w;
+	int o, r, w;
 	char *buffer;
 
 	if (filename == NULL)
